+++ Robust.Shared/ComponentTrees/TreeRecursiveMoveEvent.cs
@@ -1,0 +1,13 @@
+using Robust.Shared.GameObjects;
+
+namespace Robust.Shared.ComponentTrees;
+
+[ByRefEvent]
+internal readonly struct TreeRecursiveMoveEvent
+{
+    public readonly TransformComponent Xform;
+    public TreeRecursiveMoveEvent(TransformComponent xform)
+    {
+        Xform = xform;
+    }
+}