+++ Robust.UnitTesting/Server/RobustServerSimulation.cs
@@ -257,6 +257,7 @@ namespace Robust.UnitTesting.Server
 
             var compFactory = container.Resolve<IComponentFactory>();
 
+            // if only we had some sort of attribute for autmatically registering components.
             compFactory.RegisterClass<MetaDataComponent>();
             compFactory.RegisterClass<TransformComponent>();
             compFactory.RegisterClass<MapGridComponent>();
@@ -269,6 +270,8 @@ namespace Robust.UnitTesting.Server
             compFactory.RegisterClass<PhysicsMapComponent>();
             compFactory.RegisterClass<FixturesComponent>();
             compFactory.RegisterClass<CollisionWakeComponent>();
+            compFactory.RegisterClass<OccluderComponent>();
+            compFactory.RegisterClass<OccluderTreeComponent>();
 
             _regDelegate?.Invoke(compFactory);