+++ RELEASE-NOTES.md
@@ -35,7 +35,11 @@ END TEMPLATE-->
 
 ### Breaking changes
 
-*None yet*
+* ClientOccluderComponent has been removed & OccluderComponent component functions have been moved to the occluder system.
+* The OccluderDirectionsEvent namespace and properties have changed.
+* The rendering and occluder trees have been refactored to use generic render tree systems.
+* Several pointlight and occluder component properties now need to be set via system methods.
+
 
 ### New features